requirements.txt
requests
python-dotenv
pandas
nltk
streamlit
streamlit-autorefresh
plotly
wordcloud
matplotlib